Update Jakarta EE 11 test to use minimum Java of 17 instead of 21
- When PR #27349 was created, Jakarta EE 11 features required Java 21.
As such it set the minimum Java to be 21 for the tests added.  In
parallel, PR #27335 was being done to switch to Java 17 for all Jakarta
EE 11 features and tests.  Both were merged on the same day so the new
tests were not converted.  This PR fixes the missed tests to be Java 17
minimum.