Update features that depend on appSecurity-6.0 to be marked for beta
- Update value-add features that depend appSecurity-6.0 to be beta since
appSecurity-6.0 was moved to beta.
- Update jwtSso-1.0 EE 11 private features to depend on the noship-1.0
feature since it cannot be moved to beta / ga until there is a MP
features that supports EE 11.
- Update VisibilityTest to remove the special exception for connectors
where there was a mismatch of private and public singleton features with
the same base name.
- Add a new test to VisibilityTest to make sure singleton features with
the same base name have the same edition except for a few exceptions
that are expected where we moved editions between versions.  This helps
making sure things are put in the right edition when moving them from
noship to beta or ga.