Do not cache OpenAPI model if user code called
Do not cache the OpenAPI model if the user has registered an OASFilter
or an OASModelReader. The users code could reference anything in the
environment so we can't safely invalidate the cache.

Additionally store the Jandex index (if created) in the cache. The
Jandex index can still be used if config which does not affect indexing
changes or if the user registers an OASFilter or OASModelReader.
Creating the Jandex index is the slowest part of building the OpenAPI
model.