Fix AtomicServiceReference threading issue
SCR does not protect calls to bind methods.
This leads to some unexpected ordering issues
when AtomicServiceReference is used with dynamic,
greedy, mandatory references which can lead
to a state where the reference ends up being null.

This fixes the issue by using a PriorityQueue
to order the references according to rank so that
the order the bind methods are called do not matter.