Update to move data-1.0 feature to core edition
- Because of data-1.0's dependency on persistenceService-2.0 it had been
put into the base edition instead of the core edition.  This will be a
problem for when webProfile-11.0 is moved to beta since that feature is
in the core edition.
- As part of this change the persistenceService-2.0 and
org.eclipse.persistence 3.x features are moved back to core as well.
- Also the VersionlessTest is updated to no longer skip noship features.
This is a follow on change now that all noship features from jakartaee
11 are now referenced by the versionless features.