Issue #28911: Intermittent java:global NameNotFoundException Update injection scope code that handles deferred processing for java:global and java:app to properly handle concurrent threads looking up JNDI names in these contexts.
Previously, the first thread would begin processing the deferred metadata and the second thread
would then assume it was already complete and fail. The change will now result in the second
thread waiting on a Future for the first thread to complete.

Also updated Jakart Data code to report this failure immediately, rather than ignoring it
until the failure potentially occurs at a later time.