Fix dynamic config updates with wasJmsEndpoint
The service component did not have a modified method
which results in it getting deactivated and reactivated
on config updates.  The rest of the system does not handle
that well.

This change adds the modified method to update the inbound chain
configuration dynamically.

This change also delays the inbound port opening until after
restore for checkpoint.