Delay activation of OpenTelemetryLogHandler until restore
Enabling the log handler on the checkpoint side causes the
OTEL stack to get activated too early.  This causes issues
with reading the env configuration before restore and the
OTEL stack also needs all of the JVM security services to
startup up correctly (e.g. the SSLContext) which are not
fully available until restore.

This change adds the SatisfyingConditionTarget to the
checkoint process running condition.  This condition is not
available until the restore side which prevents this component
from getting activated until restore.