Add additional tests for FFDC with InstantOn
Also add tests that produce FFDC during startup.
For the checkpoint tests it confirms the early startup FFDCs
that get produced on the checkpoint side do not show up again
in Otel on restore.

For regular runs the early start FFDC produced by a user feature
bundle does not get properly logged to Otel.  The assert for this
is currently commented out.