Telemetry test: permission to get local host name
Give test code permission to get the real value from
InetAddress.getLocalHost so that it can correctly get the expected value
for host.name when SecurityManager is enabled.

On most systems, getLocalHost seems to return something which matches
"localhost" which can be resolved with default permissions, but on some
systems we see a different value returned from getLocalHost in app code
vs. getLocalHost in liberty code when SecurityManager is enabled.