Disable SASL Tests for Java versions >21
The Kafka Client's SASL code uses the Java security, which as of Java 23 now throws exceptions if used. there is currently no Kafka Client version that does not use Security Manager. To allow for the rest of the RM suite of tests to continue passing without the failing Java 23 build, set the maximum java level to 21.

This limitation should be removed once a suitable client is available and included.