Wait for Exception log line instead of immediately looking.
In the BadConnector test, we expect App start to fail given this is an explicit invalid context test. As such we only wait for Server start, not app start. In this case it is possible that the server is declared ready, the logs checked for the error immediately upon the server being declared up. but the application and the RM initialization has not gotten to a sufficient stage to emit the exception.

By allowing time for the exception to occur after app start as latest run had just over 20 seconds between Server start and exception being thrown, we should prevent this test from being declared a failure when it would of passed if we had been more patient.