Updates to not include more in autoFVT.zip files than necessary
- Remove things that are not needed in requiredLibs since they are
either not used or already part of the list of jars that are already
available to all fats like io.openliberty.org.apache.commons.logging
- Switch to have requiredLibs and other configs possibly to not load
transitively to not load everything that is in the buildpath for a
project.  This change prevents adding extra jars to autoFVT.zip since
they are not needed.