Fully read heartbeat stream in RESTMBeanServerConnection
The server response to the heartbeat/notification request is
the class name string 'com.ibm.ws.kernel.boot.jmx.internal.PlatformMBeanServer'.
This needs to be read, and the stream closed, to allow the HTTPURLConnection
to be properly pooled/reused. Not reading the stream was causing dangling
sockets in close_wait state