Add FIPS140-3 support for CLI commands
Add support for a environment variables that allows all CLI commands to run with the various FIPS options set. The variable can be supplied just the same as any other environment variable that is picked up for a server, client or tool. For Server and Client, The name of the profile is retrieved from the file specified, allowing for the use of a singular property. Tools generated from the tool/tool.bat script only need the property to set

Empty String e.e. ENABLE_FIPS140_3="" is set for either Windows or *nix then these will work with IBM SDK8
On Windows if JAVA_HOME is set to an Java 8 JVM, it is updated to be `!JAVA_HOME!\jre`. so the JAVA_HOME FIPS 140-3 IBM Java 8 check must not include `\jre` in the path as it is already set

This will be updated to include the Liberty baseline

The the Client and Server scripts can handle multiple files within the property for retrieving the profileName, but specifying more then one within the property will cause issues for Java itself as only one of the file specified will be laoded.