Include Liberty base FIPS140-3 profile in WLP installs
Add new task that copies the base Liberty platform into lib directory.
As this is the base Liberty profile, it is included within the Lib directory as the file is owned by the product.

Included as part of PackageLibertyWithFeatures as the new task works for wlp-embeddable and ol-dev images which include the entire `wlp` directory, but the file is not pulled in as part of the OL GA, NOSHIP or BETA Images without a copy as we do for the relevant licenses and notices files