Reset SyncToOSThread in EJB After Each Invocation
Description:

We need to update the code to reset the thread identity back to the server identity when EJBSecurityCollaboratorImpl.postInvoke() is called.

Currently, although we are correctly setting the syncToOSThreadToken in the EJBSecurityContext during preInvoke, this token is not being propagated to postInvoke because we are returning a SecurityCookieImpl object that does not contain the token field.

This fix adds the syncToOSThreadToken field to SecurityCookieImpl and ensures the token is properly transferred from the EJBSecurityContext to the cookie during preInvoke, allowing postInvoke to retrieve and use it to reset the thread identity.